+++ src/theme/dashboard-resize-handle.ts
@@ -1,0 +1,8 @@
+/**
+ * DashboardResizeHandle
+ * A resize handle for sidebars or panels.
+ * ---
+ */
+export default {
+  base: 'hidden lg:block touch-none select-none cursor-ew-resize relative before:absolute before:inset-y-0 before:-left-1.5 before:-right-1.5 before:z-1'
+}